refactor(domain,view,tui): move the isolated Model into kanban-domain
The Model was in kanban-view, which put it above kanban-service in the
dependency graph and so out of reach of the service layer. Moving it down
to kanban-domain, the lowest crate every consumer already reaches, fixes
that. Every type it is built from (LoadState, Snapshot, Board, Card,
DependencyGraph) already lives there.

The move is only legitimate once the Model stops being presentation-aware,
so the same commit strips its nine presentation fields into a new
kanban_view::controller::Controller: the four displayed_* partitions, the
archived_board_at side map and the four board-sort scalars, plus the nine
methods over them. The Model drops from twenty fields to eleven and its
imports lose kanban_core::AppConfig, chrono, filter_and_sort_boards,
BoardListFilter, BoardSortField and SortOrder. That absence is the
mechanical proof the isolation landed and is checked by a grep in the PR.

Model and Controller are peers on App, read direction Controller -> Model.
Controller::sync(&Model) is the single rebuild entry point, and every
snapshot load in kanban-tui now routes through the new App::load_snapshot
so the partitions can never lag the model. The rebuild is not per frame.

load_from_snapshot keeps its exact name and signature. apply_resolved and
mark_failed move with the Model and lose their partition rebuilds only;
neither has a caller outside apply.rs, and both now document that a future
caller must follow with Controller::sync.

Also adds a feature-gated Model::with_load_states(ModelLoadStates) so
out-of-crate tests can build a Model with arbitrary per-collection load
states. Gated on any(test, feature = "test-helpers") so a plain
cargo test -p kanban-domain does not silently skip it.

Test dispositions: twelve tests move to the Controller unchanged in their
assertions, one is deleted
(test_apply_resolved_rebuilds_the_displayed_partitions, subsumed by the
stronger test_sync_rebuilds_the_partitions_after_the_model_changes), and
one is narrowed to the index assertions that are still the Model's
(test_apply_resolved_leaves_indexes_untouched_for_an_untouched_tier).
