+++ .changeset/kan-1359-move-isolated-model-kanban-domain.md
@@ -1,0 +1,5 @@
+---
+bump: minor
+---
+
+domain,view,tui: move the `Model` out of `kanban-view` into `kanban-domain` and strip its nine presentation fields into a new `kanban_view::controller::Controller`. The `Model` now holds entities and load state only (eleven fields), so `kanban-service` will be able to name it without depending on the presentation layer. The four `displayed_*` partitions, the `archived_board_at` side map and the four board-sort scalars, together with the nine methods over them, become `Controller` state rebuilt by the new `Controller::sync(&Model)`; `App` holds the two as peers and routes every snapshot load through `App::load_snapshot`. Adds a feature-gated `Model::with_load_states(ModelLoadStates)` test constructor behind `kanban-domain`'s new `test-helpers` feature. This is a `minor` bump rather than `patch` because it is breaking under the pre-1.0 rule in `.changeset/README.md`: the public type `kanban_view::model::Model` and the public module `kanban_view::model` are removed, and nine public methods move from `Model` to a different type in a different crate. Every crate in `crates/` publishes to crates.io, so external callers of `kanban_view::model::Model` break.
