+++ crates/kanban-tui/tests/archived_board_parity_tests.rs
@@ -45,7 +45,7 @@ fn seed_and_archive_board(
     let sprint = app.ctx.create_sprint(board.id, None, None).unwrap();
     app.ctx.archive_board(board.id).unwrap();
     let snap = app.ctx.snapshot().unwrap();
-    app.model.load_from_snapshot(snap);
+    app.load_snapshot(snap);
     (board.id, col.id, card1.id, sprint.id)
 }
 
@@ -190,7 +190,7 @@ fn test_archived_board_card_action_works() {
     // Toggle completion via the SAME handler a live board's card uses.
     app.handle_toggle_card_completion();
     let snap = app.ctx.snapshot().unwrap();
-    app.model.load_from_snapshot(snap);
+    app.load_snapshot(snap);
 
     let card = app
         .model
@@ -220,7 +220,7 @@ fn test_archived_board_kanban_view_honours_board_setting() {
         )
         .unwrap();
     let snap = app.ctx.snapshot().unwrap();
-    app.model.load_from_snapshot(snap);
+    app.load_snapshot(snap);
 
     // Browsing the archived LIST is never kanban (the list must stay visible)...
     app.mode = AppMode::ArchivedBoardsView;
@@ -480,7 +480,7 @@ fn test_restore_card_reachable_from_drilled_in_archived_board() {
     app.ctx.archive_card(card.id).unwrap();
     app.ctx.archive_board(board.id).unwrap();
     let snap = app.ctx.snapshot().unwrap();
-    app.model.load_from_snapshot(snap);
+    app.load_snapshot(snap);
 
     open_archived_board(&mut app);
     app.handle_archived_boards_view_mode(crossterm::event::KeyCode::Char('D'));
